Update the repository and view model
The repository is updated to have the authentication service passed in
via the initialiser, and to retrieve the user ID from the
authentication service. The view model is updated to create the
authentication service and pass this to the repostiry when it is
created. It also retrieves the user from the authentication service and
pulsishes this.