Update card list view to show sign in screen
This is displayed if there is no user signed in. This doesn't work when
the anonymous user has been previously used to set up cards. Deleting
the users doesn't appear to correct the problem either, though they
could be cached for a while - perhaps? The result is that the sign in
screen is displayed for a short period, then a blank list is shown.
Even tried to delete the app from the phone simulator.

Running on a new simulator appears to be OK, since it doesn't know
about any users.